Remove the ‘fetch’ instruction in favour of a different approach…
Prefer a ‘get’ instruction that notionally returns
a Maybe and does not enforce invariants.

The type system can’t represent Maybes, so instead
the get instruction will return two values, one
will be the value at the array index and the
second will be whether there is no value and the
requested index was out of bounds.

The value returned when the index is out of bounds
will just be a default, e.g. false or 0.