Don’t depend on PEG.js…
There’s no need to depend on PEG.js other than in
development. This change makes it so that the unit
tests still generate the grammar on the fly, but
the integration tests use the pre-generated
grammar.

The parser rule is regenerates the grammar and is
incorporated into the build pipeline.