Make assignment destructure by default…
This is roughly based on Ruby. It means that there
is no need for special destructuring syntax. The
one downside to this approach is that expressions
have to be evaluated right-left, which could be
counter-intuitive. The only time this would matter
is if one of the expressions on the right has
side-effects. e.g.

```
x = 1;
function^ foo () { x *= 2; return 123; };
a, b = x, foo();
```

This would assign a value of 2 to a because the
dynamically scoped ‘foo’ function evaluated first
and changed the value of ‘x’.