Make recursion detection smarter…
When functions get passed around, they can take on
new names. It is no longer adequate to simply
track function names on the call stack. This makes
it so that a unique id is given to a function when
it is registered, which allows us to detect when
we are calling the same function more than once
in the call stack.