Infer Minisat Memory requirement from problem...
Minisat is originally a C program and has been compiled into
JavaScript with Emscripten. It uses native JavaScript ArrayBuffer
objects for the LLVM stack and heap. By default, it allocates a
fixed size for these data structures, but it can be configured
at compile time to automatically grow these as necessary.
Unfortunately, using this setting impacts performance.

This commit first updates Minisat to the latest version, then
wraps it in some code that lets you pass through the
TOTAL_MEMORY and TOTAL_STACK options. When running tests, we
set this once at the start of the suite to values that are
large enough to accomodate the test scenarios. We avoid
loading the Minisat module multiple times because this is slow.

When running in production, these sizes are instead inferred
from the complexity of the problem using a "number of lines"
heuristic of the provided DIMACS file. The multiplier values
were determined by debugging from inside the Minisat module to
see how much stack is allocated for different problem sizes.
These numbers should give Minisat some breathing room whilst
being conservative in its use of resources.

Here are the resource stats for different problem sizes:

100k lines: 50MB heap, 18MB stack
250k lines: 100MB heap, 33MB stack
500k lines: 168MB heap, 59MB stack
1mil lines: 319MB heap, 108MB stack
4mil lines: 1225MB heap, 410MB stack

For the last result, it was necessary to increase node's
memory limit as well, with --max-old-space-size=2000