Automatically grow heap/stack on error…
The previous heuristic strategy is good to decide
on an initial heap/stack size for Minisat, but it
falls down when sentient is run with the ‘number’
option. Eventually, Minisat runs out of either
heap or stack space.

This change makes it so that Minisat is
reinitialised with +16MB/+8MB heap/stack when it
runs out of space. This requires a few patches to
the minified vendor/minisat.js which are quite
detailed. I’ve included a vendor/README.md that
explains these steps.