Don’t exclude literals that don’t matter…
It’s incorrect to exclude literals (via adding a
clause) if those literals don’t appear in the 
level 1 variables for the problem. It would mean
that you could end up with the same solution 
multiple times if an irrelevant variable is
assigned differently.

This also massively speeds up multiple solution
runs as the written clauses are much smaller and
the SAT solver doesn’t have to churn through as
much.