Add bin/ to node module as per cdnjs setup instructions
I'm in the process of setting up Sentient on cdnjs.com.
There's a hook that adds a new version to the cdn every
time the node module is updated. That means we need to
include the bin/sentient.js file in the node package
so that it can upload it.